Course 엔티티 필드 수정 및 두루누비 코스 데이터 완성 (#89)

+++ src/main/java/com/server/running_handai/domain/course/dto/CourseDetailDto.java
@@ -7,7 +7,7 @@ import java.util.List;
 public record CourseDetailDto(
         Long courseId,
         String courseName,
-        int distance,
+        double distance,
         int duration,
         Double minElevation,
         Double maxElevation,
