+++ src/main/java/com/server/running_handai/domain/course/dto/CourseInfoDto.java
@@ -3,7 +3,7 @@ package com.server.running_handai.domain.course.dto;
 public interface CourseInfoDto {
     long getId();
     String getThumbnailUrl();
-    int getDistance();
+    double getDistance();
     int getDuration();
     double getMaxElevation();
     double getDistanceFromUser(); // 코스 시작점과 사용자의 거리
